Optimize the pod fetching log function (#469)
