+++ pkg/util/util.go
@@ -18,11 +18,12 @@ package util
 
 import (
 	"fmt"
-	"io/ioutil"
 	"math/rand"
+	"net/http"
 	"os"
 	"time"
 
+	"github.com/gorilla/websocket"
 	"golang.org/x/crypto/bcrypt"
 )
 
@@ -121,5 +122,16 @@ func EnsureDirectoryExists(path string) error {
 }
 
 func WriteToFile(filename string, data []byte) error {
-	return ioutil.WriteFile(filename, data, 0644)
+	return os.WriteFile(filename, data, 0644)
+}
+
+func BuildWebSocketConnection(w http.ResponseWriter, r *http.Request) (*websocket.Conn, error) {
+	upgrader := websocket.Upgrader{
+		ReadBufferSize:  1024,
+		WriteBufferSize: 1024,
+		CheckOrigin: func(r *http.Request) bool {
+			return true
+		}}
+	upgrader.Subprotocols = []string{r.Header.Get("Sec-WebSocket-Protocol")}
+	return upgrader.Upgrade(w, r, nil)
 }
